+++ config/splat.us.stno1.yaml
@@ -59,6 +59,7 @@ segments:
       - [0x34A00, .rodata, e_stage_name]
       - [0x34A34, .rodata, e_skeleton]
       - [0x34A50, .rodata, unk_4DDD8]
+      - [0x34B6C, .rodata, unk_51DAC]
       - [0x34CBC, .rodata, prim_helpers]
       - [0x34CC4, c, first_c_file]
       - [0x34E6C, c, e_red_door]
@@ -77,6 +78,7 @@ segments:
       - [0x4D2A8, c, popup]
       - [0x4D5D8, c, e_skeleton]
       - [0x4DDD8, c, unk_4DDD8]
+      - [0x51DAC, c, unk_51DAC]
       - [0x55AF8, c, prim_helpers]
       - [0x56328, sbss]
       # - [0x57024, .bss, e_collect]
